fix: first review batch (WIP checkpoint, squashed later)
- gpt.py: real 'adamw' optimizer branch; adamw and muon were identical
- ordinal_scheduler.py: schedule scales each param group's own base LR
- reversible_block_torch.py: recompute in backward re-enters forward autocast state
- gauge_block_torch.py: interleave via stack instead of in-place writes
- diagnostics_data.py / tests: version-bump gate hashes content, not parquet bytes
- cli.py eval: payload built for both --export-json and --artifacts-dir
- e2e_pipeline.py: scorecard contract expects manifest schema v3
- tests/__init__.py: disable Typer forced terminal (GITHUB_ACTIONS color)
- hoss_opt_torch.py: drop no-op cast; surreal_torch.py: drop dead probe
- pyproject/.beads: repo URL, br lock ignore

+++ nanochat/hoss_opt_torch.py
@@ -153,8 +153,9 @@ class HOSS(Optimizer):
 
             device = params_flat.device
 
-            # Cast to float32 for stability in Lanczos
-            params_flat.float()
+            # Lanczos runs in float32 for stability; the parameter vector itself
+            # is only needed for its device/layout (updates are cast back per
+            # parameter below).
             grads_flat_f32 = grads_flat.float()
 
             # Gradient clipping
